Remove unnecessary code
